# simplify dot example