--- a/examples/dot/asm.go
+++ b/examples/dot/asm.go
@@ -31,8 +31,8 @@ func main() {
 	blockitems := 8 * unroll
 	blocksize := 4 * blockitems
 	Label("blockloop")
-	CMPQ(n, U32(blockitems))
-	JL(LabelRef("tail"))
+	CMPQ(n, U32(0))
+	JE(LabelRef("reduce"))
 
 	// Load x.
 	xs := make([]VecVirtual, unroll)
@@ -49,28 +49,13 @@ func main() {
 		VFMADD231PS(y.Offset(32*i), xs[i], acc[i])
 	}
 
+	CMPQ(n, U32(blockitems))
+	JL(LabelRef("reduce"))
+
 	ADDQ(U32(blocksize), x.Base)
 	ADDQ(U32(blocksize), y.Base)
 	SUBQ(U32(blockitems), n)
 	JMP(LabelRef("blockloop"))
-
-	// Process any trailing entries.
-	Label("tail")
-	tail := XMM()
-	VXORPS(tail, tail, tail)
-
-	Label("tailloop")
-	CMPQ(n, U32(0))
-	JE(LabelRef("reduce"))
-
-	xt := XMM()
-	VMOVSS(x, xt)
-	VFMADD231SS(y, xt, tail)
-
-	ADDQ(U32(4), x.Base)
-	ADDQ(U32(4), y.Base)
-	DECQ(n)
-	JMP(LabelRef("tailloop"))
 
 	// Reduce the lanes to one.
 	Label("reduce")
@@ -82,7 +67,6 @@ func main() {
 	top := XMM()
 	VEXTRACTF128(U8(1), acc[0], top)
 	VADDPS(result, top, result)
-	VADDPS(result, tail, result)
 	VHADDPS(result, result, result)
 	VHADDPS(result, result, result)
 	Store(result, ReturnIndex(0))