Thanks for avo, it's a great project!

I noticed that it's possible to simply remove the tail part of SIMD programs and the resulting assembly runs without issues and passes the tests. Avoiding the tailloop makes writing SIMD programs significantly easier and simpler (although in this example it's a relatively minor change). Additionally, the simplification makes the operation ~2.5x faster for inputs of size 512.

This PR is partially a genuine simplification and partially a question of why this doesn't crash?

Notably, I expected 
```
VMOVUPS(x.Offset(32*i), xs[i])
```
to fail, if `x.Offset(32*i)` is out of bounds. But, maybe this actually copies no data and therefore the program doesn't crash?

Using:
```
func BenchmarkDot(b *testing.B) {
	n := 512
	x, y := RandomVector(n), RandomVector(n)
	for i := 0; i < b.N; i++ {
		Dot(x, y)
	}
}
```
It seems the performance is significantly in favor of the simplification:

Current:
```
goos: linux
goarch: amd64
pkg: github.com/mmcloughlin/avo/examples/dot
cpu: AMD Ryzen 5 3600 6-Core Processor              
BenchmarkDot-12    	27368935	       44.80 ns/op
PASS
ok  	github.com/mmcloughlin/avo/examples/dot	1.289s
```

With PR applied:
```
$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/mmcloughlin/avo/examples/dot
cpu: AMD Ryzen 5 3600 6-Core Processor              
BenchmarkDot-12    	67842052	       18.01 ns/op
PASS
ok  	github.com/mmcloughlin/avo/examples/dot	1.255s
```