@y0ast 2021-04-07
Turns out that more proper testing (by using padded slices) shows that this can obviously lead to wrong answers (just not crashes). This fails for `n = 1` already:

```
func TestLengths(t *testing.T) {
	const epsilon = 0.00001
	for n := 0; n < 1000; n++ {
		x, y := RandomVector(n+50), RandomVector(n+50)

		got := Dot(x[25:len(x)-25], y[25:len(y)-25])
		expect := Expect(x[25:len(x)-25], y[25:len(y)-25])
		relerr := got/expect - 1.0
		if Abs(relerr) > epsilon {
			t.Fatalf("bad result on vector length %d: got %v expect %v relative error %f", n, got, expect, relerr)
		}
	}
}
```