@mmcloughlin 2021-04-11
Hey there @y0ast sorry for not replying sooner!

Seems you figured it out but I'm confident that the tail logic is necessary, though I agree it's annoying and produces most of the complexity. It is interesting that when you removed it you didn't get a segfault or the wrong answer, I would guess this is due to internal details of the Go memory allocator.